fix: config test
